+++ main-app/client/src/assets/scss/Notifications.scss
@@ -84,7 +84,7 @@
         padding: 15px 10px;
         border-radius: 15px;
         box-shadow: rgba(50, 50, 93, 0.25) 0px 6px 12px -2px, rgba(0, 0, 0, 0.3) 0px 3px 7px -3px;
-        transition: all 0.5s ease;
+        transition: all 0.2s ease;
         &:hover{
             border-left: 40px #294A70 solid;
         }
