+++ main-app/client/src/pages/Admin/requestor.js
@@ -1,9 +1,9 @@
-import TableHead from './TableHeading';
-import UserTable from './TableContent';
-import '../assets/scss/user-management.scss';
-import Sidenav_Admin from './Sidenav_Admin';
+import TableHead from '../TableHeading';
+import UserTable from '../TableContent';
+import '../../assets/scss/user-management.scss';
+import Sidenav_Admin from '../Admin/Sidenav_Admin';
 
-function Archived() {
+function RequestorManagement() {
     return (
         <div className="app">
         <div id="side">
@@ -12,10 +12,8 @@ function Archived() {
         <div id="home">
         <div class='container'>
             <div class='title'>
-                ARCHIVED <div class='yellow'>MANAGEMENT</div>
+                REQUESTOR <div class='yellow'>ACCOUNTS</div>
             </div>
-
-            
 
             <div class='search'>
                 <input placeholder='Search By ID'></input>
@@ -23,18 +21,19 @@ function Archived() {
             </div>
 
             <div class='table-heading'>
-                <TableHead />
+                <TableHead/>
             </div>
             <div class='table-cell'>
-                <UserTable />
-                <UserTable />
-                <UserTable />
-                <UserTable />
-                <UserTable />
+                <UserTable/>
+                <UserTable/>
+                <UserTable/>
+                <UserTable/>
+                <UserTable/>
             </div>
             <div class='buttons'>
-                <a>RESTORE USER</a>
-                <a>DELETE</a>
+                <div>ARCHIVE</div>
+                <div>UPDATE USER</div>
+                <div>ADD USER</div>
             </div>
         </div>
         </div>
@@ -43,4 +42,5 @@ function Archived() {
     );
 }
 
-export default Archived;
+
+export default RequestorManagement;
