+++ main-app/client/src/pages/Admin/user-req-table.js
@@ -1,6 +1,6 @@
-import TableHead from './TableHeading';
-import UserTable from './TableContent';
-import '../assets/scss/Table.scss';
+import TableHead from '../TableHeading';
+import UserTable from '../TableContent';
+import '../../assets/scss/Table.scss';
 import Sidenav_Admin from './Sidenav_Admin';
